// calorie-counter-app/src/typings/foodCondimentFacts.ts
export interface FoodCondimentFacts {
  survey_food_code: number;
  display_name: string;
  condiment_portion_size: string;
  condiment_portion_code: number;
  condiment_grains: string;
  condiment_whole_grains: string;
  condiment_vegetables: string;
  condiment_dkgreen: string;
  condiment_orange: string;
  condiment_starchy_vegetables: string;
  condiment_other_vegetables: string;
  condiment_fruits: string;
  condiment_milk: string;
  condiment_meat: string;
  condiment_soy: string;
  condiment_drybeans_peas: string;
  condiment_oils: string;
  condiment_solid_fats: string;
  condiment_added_sugars: string;
  condiment_alcohol: string;
  condiment_calories: string;
  condiment_saturated_fats: string;
}